MakeMaker kind of requires two separate inc/ copies
Due to how MakeMaker works, we need to maintain separate copies of the
code in inc/ (which is required for Makefile.PL). Symlinks fuck of the
make dist and make manifest stages of distribution building and we
really don't want that.

Yves, if this breaks your build again, try a fresh clone!