add tests and fix round trips of utf8 keys in the latin1 range
Perl stores hashes in two forms: UTF8 and Latin-1 (binary).

If a unicode key is stored in a hash then perl will see if it can be reencoded as
latin-1 (binary). If it can then it is stored as latin-1 with the flag "WASUTF8".
If it cant then it is stored as utf8 with the flag ISUTF8.

If the key is latin-1/binary from the beginning then it is stored as latin-1 with
neither flag set.

When fetching the keys of a hash and the key has the WASUTF8 then perl upgrades
the string before returning it to the user so it has the correct semantics. This
means that we also have to do the same thing for WASUTF8 keys or they will not
round trip properly.