Do not die on undef PVMG's
We end up dieing because the SV we are dumping is a magic SV, which
we have called mg_get() on, which has been legitimately set to undef.

I am not sure if there is a better way to detect whether PVMGs have been
handled ok. We could maybe rework this logic further, but it seems ok
now.