Perl: "Sereal" meta distribution
The "Sereal" module/distribution depends on both encoder and decoder for
convenience. It will always depend on the most recent version of both
that have a lower or equal version as the Sereal meta-distribution
itself.

This is not Task::Sereal since it actually offers Sereal encoder and
decoder functions for exporting. (Tasks are just dependency vehicles.)