Old-perl compatibility fix
This (hopefully) fixes one of the failures with older Perls: The
MUTABLE_* macros weren't present. Their sole purpose is to warn about
inadvertently casting away constness, but since we know our input isn't
const, that's kind of a silly thing to do all over the place anyway?
I guess I'm missing some detail with this, but it's quite clear that
using that macro in the particular place doesn't actually do anything
meaningful, so nuke it for old times sake.